Paginação, Ordenação e Cache

+++ src/main/java/br/com/ghabriel/forum/controller/TopicosController.java
@@ -1,13 +1,17 @@
 package br.com.ghabriel.forum.controller;
 
 import java.net.URI;
-import java.util.List;
 import java.util.Optional;
-
 import javax.transaction.Transactional;
-import javax.validation.Valid;
+
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort.Direction;
+import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -16,6 +20,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.util.UriComponentsBuilder;
 
@@ -38,12 +43,14 @@ public class TopicosController {
 	private CursoRepository cursoRepository;
 
 	@GetMapping
-	public List<TopicoDto> lista(String nomeCurso) {
+	@Cacheable(value = "listaDeTopicos")
+	public Page<TopicoDto> lista(@RequestParam(required = false) String nomeCurso, @PageableDefault(sort = "id", direction = Direction.ASC, page = 0, size = 10) Pageable paginacao) {
+
 		if (nomeCurso == null) {
-			List<Topico> topicos = topicoRepository.findAll();
+			Page<Topico> topicos = topicoRepository.findAll(paginacao);
 			return TopicoDto.converter(topicos);
 		} else {
-			List<Topico> topicos = topicoRepository.findByCurso_Nome(nomeCurso);
+			Page<Topico> topicos = topicoRepository.findByCurso_Nome(nomeCurso, paginacao);
 			return TopicoDto.converter(topicos);
 		}
 
@@ -51,13 +58,14 @@ public class TopicosController {
 
 	@PostMapping
 	@Transactional
+	@CacheEvict(value = "listaDeTopicos")
 	public ResponseEntity<TopicoDto> cadastrar(@RequestBody TopicoForm form, UriComponentsBuilder uriBuilder)
 			throws Exception {
 
 		if (form.getTitulo() == null || form.getMensagem() == null || form.getNomeCurso() == null
 				|| form.getTitulo().isEmpty() || form.getMensagem().isEmpty() || form.getNomeCurso().isEmpty()) {
 
-			throw new Exception("Campos inválidos");
+			return ResponseEntity.notFound().build();
 
 		} else {
 			Topico topico = form.converter(cursoRepository);
@@ -71,14 +79,18 @@ public class TopicosController {
 	}
 
 	@GetMapping("/{id}")
-	public DetalhesDoTopicoDto detalhar(@PathVariable Long id) {
-		Topico topico = topicoRepository.getOne(id);
-		return new DetalhesDoTopicoDto(topico);
+	public ResponseEntity<DetalhesDoTopicoDto> detalhar(@PathVariable Long id) {
+		Optional<Topico> topico = topicoRepository.findById(id);
+		if (topico.isPresent()) {
+			return ResponseEntity.ok(new DetalhesDoTopicoDto(topico.get()));
+		}
+		return ResponseEntity.notFound().build();
 
 	}
 
 	@PutMapping("/{id}")
 	@Transactional
+	@CacheEvict(value = "listaDeTopicos")
 	public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody AtualizacaoTopicoForm form)
 			throws Exception {
 		if (form.getTitulo() == null || form.getMensagem() == null || form.getTitulo().isEmpty()
@@ -87,17 +99,28 @@ public class TopicosController {
 			throw new Exception("Campos inválidos");
 
 		} else {
-			Topico topico = form.atualizar(id, topicoRepository);
-			return ResponseEntity.ok(new TopicoDto(topico));
+
+			Optional<Topico> optional = topicoRepository.findById(id);
+			if (optional.isPresent()) {
+				Topico topico = form.atualizar(id, topicoRepository);
+				return ResponseEntity.ok(new TopicoDto(topico));
+			}
+			return ResponseEntity.notFound().build();
 
 		}
 	}
-	
+
 	@DeleteMapping("/{id}")
 	@Transactional
-	public ResponseEntity<?> remover(@PathVariable Long id){
-		topicoRepository.deleteById(id);
-		return ResponseEntity.ok().build();
+	@CacheEvict(value = "listaDeTopicos")
+	public ResponseEntity<?> remover(@PathVariable Long id) {
+
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if (optional.isPresent()) {
+			topicoRepository.deleteById(id);
+			return ResponseEntity.ok().build();
+		}
+		return ResponseEntity.notFound().build();
 
 	}
